requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-ckeditor-5==0.0.15
django-crispy-forms==1.14.0
django-filter==21.1
django-js-asset==2.0.0
django-taggit==3.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
Markdown==3.3.7
Pillow==9.0.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
